Moved STO validator from schema to new file validators.py
