+++ bpx/__init__.py
@@ -6,6 +6,7 @@ __version__ = "0.3.1"
 from .interpolated_table import InterpolatedTable
 from .expression_parser import ExpressionParser
 from .function import Function
+from .validators import check_sto_limits
 from .schema import BPX
 from .parsers import parse_bpx_str, parse_bpx_obj, parse_bpx_file
 from .utilities import get_electrode_stoichiometries, get_electrode_concentrations
