+++ bpx/schema.py
@@ -2,7 +2,7 @@ from typing import List, Literal, Union, Dict
 
 from pydantic import BaseModel, Field, Extra, root_validator
 
-from bpx import Function, InterpolatedTable
+from bpx import Function, InterpolatedTable, check_sto_limits
 
 from warnings import warn
 
@@ -294,45 +294,6 @@ class Experiment(ExtraBaseModel):
         example=[298, 298, 298, 298, 298],
         description="Temperature vs time",
     )
-
-
-# Validates that the STO limits subbed into the OCPs give the correct voltage limits.
-# Works if both OCPs are defined as functions.
-# Blended electrodes are not supported.
-# This is a reusable validator to be used for both DFN/SPMe and SPM parameter sets.
-# https://docs.pydantic.dev/latest/usage/validators/#root-validators
-def check_sto_limits(cls, values):
-    try:
-        ocp_n = values.get("negative_electrode").ocp.to_python_function()
-        ocp_p = values.get("positive_electrode").ocp.to_python_function()
-    except AttributeError:
-        # OCPs defined as interpolated tables; do nothing
-        return values
-
-    sto_n_min = values.get("negative_electrode").minimum_stoichiometry
-    sto_n_max = values.get("negative_electrode").maximum_stoichiometry
-    sto_p_min = values.get("positive_electrode").minimum_stoichiometry
-    sto_p_max = values.get("positive_electrode").maximum_stoichiometry
-    V_min = values.get("cell").lower_voltage_cutoff
-    V_max = values.get("cell").upper_voltage_cutoff
-
-    # Checks the maximum voltage estimated from STO
-    V_max_sto = ocp_p(sto_p_min) - ocp_n(sto_n_max)
-    if V_max_sto > V_max:
-        warn(
-            f"The maximum voltage computed from the STO limits ({V_max_sto} V) "
-            f"is higher than the maximum allowed voltage ({V_max} V)"
-        )
-
-    # Checks the minimum voltage estimated from STO
-    V_min_sto = ocp_p(sto_p_max) - ocp_n(sto_n_min)
-    if V_min_sto < V_min:
-        warn(
-            f"The minimum voltage computed from the STO limits ({V_min_sto} V) "
-            f"is lower than the minimum allowed voltage ({V_min} V)"
-        )
-
-    return values
 
 
 class Parameterisation(ExtraBaseModel):
